custom command moved to api which now holds logic
